:sparkles: feedback 적용 및 수정

+++ src/main/java/com/cheeup/domain/portfolio/PortfolioDisability.java
@@ -1,8 +1,24 @@
 package com.cheeup.domain.portfolio;
 
 import com.cheeup.domain.enums.DisabilityType;
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.ConstraintMode;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.ForeignKey;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
@@ -26,6 +42,6 @@ public class PortfolioDisability {
     @Enumerated(EnumType.STRING)
     private DisabilityType type;
 
-    private Integer grade;
+    private Short grade;
 }
